New feature - downvote clocked in bad habits.
Specify bad habits as an alist where each element is of the
form (HABIT . TIME), where TIME is a string specifying when to start
downvoting the task ("1 hour", "10 minute"). The status buffer will
show a header line if we are clocked into a bad
habit. `hrpg-repeat-interval` specifies the how frequently to downvote.